Replace ENV VAR 'MASTER_TAG' by 'TAG' in build sh scripts
This replaces the hard coded var 'MASTER_TAG' by the env var 'TAG'
in the build bash scripts. So the tag for the Docker images can be
modified from outside and the 'TAG' var used in the docker-stack.yml can
be reused.
The default is 'master' as it was the hard coded value before,
so no functional change at all.